+++ CHANGELOG.md
@@ -8,6 +8,11 @@
 
 ## v0.5
 
+### v0.5.0-11
+- 🟥 `conduit.recv` now requires a 2nd argument `event`
+  - is type `ChannelMessage` (requires `data` and `origin`)
+  - this allows us to manually jack in and snoop for the `origin` on incoming messages
+
 ### v0.5.0-10
 - 🟥 WindowConduit now takes a localWindow vs a targetWindow, as apparently we listen for messages on our own window, and post to another